Extract MP4 cover art parsing into Mp4CoverExtractor; replace Long.MAX_VALUE/2 with named constant; use US_ASCII for FourCC
